+++ internal/pkg/url/deliver/url.go
@@ -7,8 +7,8 @@ import (
 	"strconv"
 
 	"github.com/labstack/echo/v4"
-	"github.com/xfiendx4life/gb_go_backend1/pkg/models"
-	"github.com/xfiendx4life/gb_go_backend1/pkg/url"
+	"github.com/xfiendx4life/gb_go_backend1/internal/pkg/models"
+	"github.com/xfiendx4life/gb_go_backend1/internal/pkg/url"
 	"go.uber.org/zap"
 )
