+++ internal/pkg/user/usecase/usecase_test.go
@@ -2,15 +2,22 @@ package usecase_test
 
 import (
 	"context"
-	"fmt"
+	"crypto/md5"
+	"encoding/hex"
+	"errors"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/xfiendx4life/gb_go_backend1/internal/logger"
-	"github.com/xfiendx4life/gb_go_backend1/pkg/models"
-	"github.com/xfiendx4life/gb_go_backend1/pkg/url/usecase"
+	"github.com/xfiendx4life/gb_go_backend1/internal/pkg/models"
+	"github.com/xfiendx4life/gb_go_backend1/internal/pkg/user/usecase"
 	"go.uber.org/zap"
 	"go.uber.org/zap/zapcore"
+)
+
+var (
+	lgr = logger.InitLogger(zapcore.DebugLevel, "")
+	ctx = context.Background()
 )
 
 type mockStorage struct {
@@ -27,33 +34,25 @@ func (mc *mockStorage) AddUser(ctx context.Context, user *models.User, z *zap.Su
 }
 
 func (mc *mockStorage) GetUserByLogin(ctx context.Context, login string, z *zap.SugaredLogger) (*models.User, error) {
+	s := md5.Sum([]byte("correctPassword"))
 	return &models.User{
 		Name:     "testname",
-		Password: "correctPassword",
+		Password: hex.EncodeToString(s[:]),
 	}, mc.err
 }
 
 func (mc *mockStorage) AddUrl(ctx context.Context, url *models.Url, z *zap.SugaredLogger) error {
-	url.Id = 1
-	return mc.err
+	return nil
 }
-
 func (mc *mockStorage) GetUrl(ctx context.Context, id int, z *zap.SugaredLogger) (*models.Url, error) {
-	return &models.Url{
-		Raw:       "RawTestUrl",
-		Shortened: "shortenedTestUrl",
-	}, mc.err
+	return nil, nil
 }
-
 func (mc *mockStorage) GetUrls(ctx context.Context, userID int, z *zap.SugaredLogger) ([]models.Url, error) {
 	return nil, nil
 }
 
 func (mc *mockStorage) GetUrlByShortened(ctx context.Context, shortened string, z *zap.SugaredLogger) (*models.Url, error) {
-	return &models.Url{
-		Raw:       "RawTestUrl",
-		Shortened: "shortenedTestUrl",
-	}, mc.err
+	return &models.Url{}, nil
 }
 
 func (mc *mockStorage) AddRedirect(ctx context.Context, r *models.Redirects, z *zap.SugaredLogger) error {
@@ -64,54 +63,44 @@ func (mc *mockStorage) GetRedirects(ctx context.Context, urlId int, z *zap.Sugar
 	return []models.Redirects{}, nil
 }
 
-var (
-	lgr = logger.InitLogger(zapcore.DebugLevel, "")
-	ctx = context.Background()
-)
-
-func TestAddUrl(t *testing.T) {
-	st := mockStorage{err: nil}
-	uc := usecase.New(&st)
-	short, err := uc.Add(ctx, "testurl", 1, lgr)
-	assert.NoError(t, err)
-	assert.NotEmpty(t, short)
-}
-
-func TestAddStorageError(t *testing.T) {
-	st := mockStorage{err: fmt.Errorf("some storage error")}
-	uc := usecase.New(&st)
-	short, err := uc.Add(ctx, "testurl", 1, lgr)
-	assert.Error(t, err)
-	assert.Empty(t, short)
-}
-
-func TestNewUser(t *testing.T) {
-	set := make(map[string]struct{})
-	var err error
-	var i int
-	var u *models.Url
-	for i = 0; i < 10000; i++ {
-		u = usecase.NewUrl("TestNewUser", 1, lgr)
-		if _, ok := set[u.Shortened]; ok {
-			err = fmt.Errorf("already exists")
-			break
-		}
-		set[u.Shortened] = struct{}{}
-	}
-	assert.NoError(t, err, fmt.Sprintf("Url: %s, i: %d", u.Shortened, i))
-
-}
-
-func TestGetNoErr(t *testing.T) {
+func TestValidateCorrect(t *testing.T) {
 	uc := usecase.New(&mockStorage{})
-	raw, err := uc.Get(ctx, "shortenedTestUrl", lgr)
+	res, err := uc.Validate(ctx, "testname", "correctPassword", lgr)
 	assert.NoError(t, err)
-	assert.Equal(t, "RawTestUrl", raw)
+	assert.True(t, res)
 }
 
-func TestGetErr(t *testing.T) {
-	uc := usecase.New(&mockStorage{err: fmt.Errorf("can't add error")})
-	raw, err := uc.Get(ctx, "SomeNotValidURL", lgr)
+func TestValidateIncorrect(t *testing.T) {
+	uc := usecase.New(&mockStorage{})
+	res, err := uc.Validate(ctx, "testname", "incorrectPassword", lgr)
+	assert.NoError(t, err)
+	assert.False(t, res)
+}
+
+func TestValidateError(t *testing.T) {
+	st := &mockStorage{err: errors.New("some error")}
+	uc := usecase.New(st)
+	res, err := uc.Validate(ctx, "testname", "incorrectPassword", lgr)
 	assert.Error(t, err)
-	assert.Equal(t, "", raw)
+	assert.False(t, res)
+}
+
+func TestAddUser(t *testing.T) {
+	uc := usecase.New(&mockStorage{})
+	u := &models.User{Name: "testname", Password: "password", Email: "email"}
+	err := uc.Add(ctx, u, lgr)
+	assert.NoError(t, err)
+	assert.Equal(t, models.User{
+		Id:       1,
+		Name:     "testname",
+		Password: "5f4dcc3b5aa765d61d8327deb882cf99",
+		Email:    "email"}, *u)
+}
+
+func TestAddUserError(t *testing.T) {
+	st := &mockStorage{err: errors.New("some error")}
+	uc := usecase.New(st)
+	u := &models.User{Name: "testname", Password: "password", Email: "email"}
+	err := uc.Add(ctx, u, lgr)
+	assert.Error(t, err)
 }
