+++ internal/pkg/user/usecase/user.go
@@ -6,8 +6,8 @@ import (
 	"encoding/hex"
 	"fmt"
 
-	"github.com/xfiendx4life/gb_go_backend1/pkg/models"
-	"github.com/xfiendx4life/gb_go_backend1/pkg/user"
+	"github.com/xfiendx4life/gb_go_backend1/internal/pkg/models"
+	"github.com/xfiendx4life/gb_go_backend1/internal/pkg/user"
 	"go.uber.org/zap"
 
 	"github.com/xfiendx4life/gb_go_backend1/storage"
